tools/create_log_file/requirements.txt
playwright==1.44.0
python-dotenv==1.0.1
